fix upgrade when starting from empty db
If you configure the db connection but don't actually execute the creation
script, the tables won't exist.
But the db migration system will think that it must execute the 1st migration
i.e. to1.0.sql, assuming that the initial creation was done.

So the idea is that if we start from 0.0, just re-execute the creation script.
And I made it idempotent, so that re-executing it would not cause an issue.