migrate db to version 3.0 to fix import issues
First, some students have long names...

Then the memo field is not mentioned in the insert query of the import code.
But the it's marked as NOT NULL without a default.
So the insert query fails.
Why wasn't it failing before? Different db setting I guess.
(see SQL_MODE).
Nevertheless, it's better to make the field optional with a sensible default.