use mysql5.5 instead of mariadb 10.5
It solves a couple of issues.
First, when inserting a record, it's allowed to not specify a field
defined as not null but without a default value.

Secondly, it can solve some issues with special characters in names.
It's allowed to insert latin1 encoded chars in utf8 defined columns.
Then it's possible to retrieve those bytes and redisplay them as latin1 chars.